# trailmemo/docs/swagger.yaml
basePath: /
definitions:
  github_com_trailmemo_internal_agent_dto.ArtifactCommitRequest:
    properties:
      commit_type:
        description: '"create_route" / "create_post"'
        type: string
      idempotency_key:
        description: 幂等键
        type: string
      is_public:
        description: 是否公开（仅 create_route）
        type: integer
    required:
    - commit_type
    - idempotency_key
    type: object
  github_com_trailmemo_internal_agent_dto.ArtifactCommitResponse:
    properties:
      artifact_id:
        description: 产物 ID
        type: string
      entity_id:
        description: 业务实体 ID
        type: integer
      entity_type:
        description: route / post
        type: string
      status:
        description: committed
        type: string
    type: object
...
  github_com_trailmemo_internal_agent_dto.ChatLoopResponse:
    properties:
      content:
        type: string
      finish_reason:
        type: string
      latency_ms:
        type: integer
      steps:
        type: integer
      tool_calls:
        type: integer
      total_tokens:
        type: integer
    type: object
  github_com_trailmemo_internal_agent_dto.ChatRequest:
    properties:
      message:
        type: string
      session_id:
        description: 可选，用于多轮对话
        type: string
    required:
    - message
    type: object
  github_com_trailmemo_internal_agent_dto.CheckpointDraftData:
    properties:
      address:
        type: string
      arrive_time:
        type: string
      city:
        type: string
      description:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      sequence:
        type: integer
      stay_duration:
        type: integer
    type: object
...
  github_com_trailmemo_internal_agent_dto.RecommendItem:
    properties:
      city:
        type: string
      days:
        type: integer
      estimated_budget:
        type: integer
      reason:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
    type: object
  github_com_trailmemo_internal_agent_dto.RecommendRequest:
    properties:
      budget:
        type: integer
      days:
        type: integer
      interests:
        items:
          type: string
        type: array
      query:
        type: string
      travel_type:
        type: string
    required:
    - query
    type: object
  github_com_trailmemo_internal_agent_dto.RecommendResponse:
    properties:
      artifact_id:
        type: string
      fallback:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.RecommendItem'
        type: array
      run_id:
        type: string
      warnings:
        items:
          type: string
        type: array
    type: object
  github_com_trailmemo_internal_agent_dto.RemixChangeItem:
    properties:
      action:
        description: added/removed/modified
        type: string
      point:
        description: 打卡点名称
        type: string
      reason:
        description: 变更原因
        type: string
    type: object
  github_com_trailmemo_internal_agent_dto.RemixRequest:
    properties:
      budget:
        type: integer
      days:
        type: integer
      query:
        type: string
      travel_styles:
        items:
          type: string
        type: array
    required:
    - query
    type: object
  github_com_trailmemo_internal_agent_dto.RemixResponse:
    properties:
      approval_required:
        type: boolean
      artifact_id:
        type: string
      change_summary:
        items:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.RemixChangeItem'
        type: array
      next_action:
        type: string
      route_draft:
        $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.RouteDraftData'
      run_id:
        type: string
      source_route_id:
        type: integer
      warnings:
        items:
          type: string
        type: array
    type: object
  github_com_trailmemo_internal_agent_dto.RenameSessionRequest:
    properties:
      title:
        type: string
    required:
    - title
    type: object
  github_com_trailmemo_internal_agent_dto.RouteDraftData:
    properties:
      checkpoints:
        items:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.CheckpointDraftData'
        type: array
      end_city:
        type: string
      estimated_budget:
        type: integer
      estimated_hours:
        type: integer
      start_city:
        type: string
      summary:
        type: string
      title:
        type: string
    type: object
  github_com_trailmemo_internal_agent_dto.RouteDraftRequest:
    properties:
      budget:
        description: 预算（元）
        type: integer
      days:
        description: 出行天数
        type: integer
      query:
        description: 用户自然语言需求
        type: string
      start_city:
        description: 出发城市
        type: string
      target_city:
        description: 目标城市
        type: string
      travel_styles:
        description: 旅行风格标签
        items:
          type: string
        type: array
    required:
    - query
    type: object
  github_com_trailmemo_internal_agent_dto.RouteDraftResponse:
    properties:
      approval_required:
        description: 是否需要确认
        type: boolean
      artifact_id:
        description: 产物 ID
        type: string
      next_action:
        description: 后续操作
        type: string
      route_draft:
        allOf:
        - $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.RouteDraftData'
        description: 路线草稿数据
      run_id:
        description: Agent 运行 ID
        type: string
      warnings:
        description: 警告
        items:
          type: string
        type: array
    type: object
  github_com_trailmemo_internal_agent_dto.SessionDetailResponse:
    properties:
      expired:
        type: boolean
      message_count:
        type: integer
      messages:
        items:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.SessionMsg'
        type: array
      session_id:
        type: string
      title:
        type: string
    type: object
...
  github_com_trailmemo_internal_agent_dto.SessionMsg:
    properties:
      content:
        type: string
      role:
        type: string
    type: object
  github_com_trailmemo_internal_agent_dto.TravelNoteRequest:
    properties:
      include_checkin_content:
        type: boolean
      include_images:
        type: boolean
      route_id:
        type: integer
      style:
        description: story/journal/guide/social/poetic
        type: string
    required:
    - route_id
    type: object
  github_com_trailmemo_internal_agent_dto.TravelNoteResponse:
    properties:
      approval_required:
        type: boolean
      artifact_id:
        type: string
      content:
        type: string
      image_refs:
        items:
          type: string
        type: array
      next_action:
        type: string
      run_id:
        type: string
      suggested_tags:
        items:
          type: string
        type: array
      title:
        type: string
      warnings:
        items:
          type: string
        type: array
    type: object
  github_com_trailmemo_internal_agent_memory.PreferenceUpdate:
    properties:
      budget_range:
        type: string
      companion_types:
        items:
          type: string
        type: array
      interests:
        items:
          type: string
        type: array
      pace:
        type: string
      preferred_cities:
        items:
          type: string
        type: array
      travel_type:
        type: string
    type: object
...
  github_com_trailmemo_pkg_response.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      error_code:
        example: ROUTE_NOT_FOUND
        type: string
      message:
        example: success
        type: string
      request_id:
        type: string
    type: object
...
host: localhost:8087
info:
  contact:
    email: support@example.com
    name: API Support
  description: 迹忆旅图寻迹旅游后端 API 文档
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://example.com/terms
  title: TrailMemo API
  version: 1.0.0
paths:
  /agent/artifacts/{artifact_id}/commit:
    post:
      consumes:
      - application/json
      description: 用户确认后将草稿导入为真实Route或Post，支持幂等
      parameters:
      - description: 产物ID
        in: path
        name: artifact_id
        required: true
        type: string
      - description: 提交请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.ArtifactCommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.ArtifactCommitResponse'
              type: object
      summary: 提交产物到业务实体
      tags:
      - 产物
...
  /agent/chat:
    post:
      consumes:
      - application/json
      description: 有界推理循环处理用户消息，意图自动路由
      parameters:
      - description: 对话请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.ChatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.ChatLoopResponse'
              type: object
      summary: Agent 对话（非流式）
      tags:
      - 对话
  /agent/chat/stream:
    post:
      consumes:
      - application/json
      description: 通过 Server-Sent Events 实时推送 AI 回复 token
      parameters:
      - description: 对话请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.ChatRequest'
      produces:
      - text/event-stream
      responses: {}
      summary: Agent 流式对话（SSE）
      tags:
      - 对话
...
  /agent/notes/generate:
    post:
      consumes:
      - application/json
      description: 从路线打卡记录自动生成游记草稿
      parameters:
      - description: 游记请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.TravelNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.TravelNoteResponse'
              type: object
      summary: 打卡生成游记
      tags:
      - 游记
  /agent/preferences:
    get:
      description: 返回AI从用户行为中学习到的偏好快照
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      summary: 查看AI画像
      tags:
      - 偏好
    put:
      consumes:
      - application/json
      description: 用户显式设置旅行偏好
      parameters:
      - description: 偏好更新
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_memory.PreferenceUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      summary: 手动更新偏好
      tags:
      - 偏好
...
  /agent/recommend:
    post:
      consumes:
      - application/json
      description: 根据用户需求和历史偏好生成目的地推荐
      parameters:
      - description: 推荐请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.RecommendRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.RecommendResponse'
              type: object
      summary: 旅行推荐
      tags:
      - 推荐
  /agent/routes/{id}/remix:
    post:
      consumes:
      - application/json
      description: 基于公开路线按需求改造（亲子版/情侣版/美食版）
      parameters:
      - description: 原路线ID
        in: path
        name: id
        required: true
        type: integer
      - description: 改造需求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.RemixRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.RemixResponse'
              type: object
      summary: 改造公开路线
      tags:
      - 路线
  /agent/routes/draft:
    post:
      consumes:
      - application/json
      description: 自然语言→结构化路线草稿（打卡点/预算/时长）
      parameters:
      - description: 路线需求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.RouteDraftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.RouteDraftResponse'
              type: object
      summary: 一句话生成路线草稿
      tags:
      - 路线
...
  /agent/sessions/{id}:
    delete:
      description: 同时清理 MySQL 元数据和 Redis 消息历史
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      summary: 删除会话
      tags:
      - 会话
    get:
      description: 获取指定会话的消息历史，含expired标记
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.SessionDetailResponse'
              type: object
      summary: 会话详情
      tags:
      - 会话
  /agent/sessions/{id}/title:
    put:
      consumes:
      - application/json
      description: 修改会话标题
      parameters:
      - description: 会话ID
        in: path
        name: id
        required: true
        type: string
      - description: 新标题
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_trailmemo_internal_agent_dto.RenameSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_trailmemo_pkg_response.Response'
      summary: 重命名会话
      tags:
      - 会话
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"